Refactor model-related code into the model/ folder
