+++ .eslintignore
@@ -1858,13 +1858,14 @@ packages/tools/checkLibPaths.test.js
 packages/tools/checkLibPaths.js
 packages/tools/convertThemesToCss.js
 packages/tools/fuzzer/ActionRunner.js
-packages/tools/fuzzer/ActionTracker.js
 packages/tools/fuzzer/Client.js
 packages/tools/fuzzer/ClientPool.js
 packages/tools/fuzzer/Server.js
 packages/tools/fuzzer/constants.js
+packages/tools/fuzzer/model/ActionTracker.js
 packages/tools/fuzzer/model/FolderRecord.js
 packages/tools/fuzzer/model/ResourceRecord.js
+packages/tools/fuzzer/model/types.js
 packages/tools/fuzzer/sync-fuzzer.js
 packages/tools/fuzzer/types.js
 packages/tools/fuzzer/utils/ProgressBar.js
