+++ packages/tools/fuzzer/ActionRunner.ts
@@ -1,6 +1,7 @@
 import Client from './Client';
 import ClientPool from './ClientPool';
-import { assertIsFolder, assertIsNote, FuzzContext, ItemId, RandomFolderOptions, ResourceData } from './types';
+import { FuzzContext, RandomFolderOptions } from './types';
+import { assertIsFolder, assertIsNote, ItemId, ResourceData } from './model/types';
 import { strict as assert } from 'assert';
 import Logger from '@joplin/utils/Logger';
 import retryWithCount from './utils/retryWithCount';
