+++ packages/tools/fuzzer/model/ActionTracker.ts
@@ -1,8 +1,9 @@
 import { strict as assert } from 'assert';
-import { ActionableClient, FolderData, FuzzContext, ItemId, NoteData, ShareOptions, TreeItem, assertIsFolder, isFolder, isNote, isResource } from './types';
-import FolderRecord from './model/FolderRecord';
+import { ActionableClient, FuzzContext, ShareOptions } from '../types';
+import { FolderData, ItemId, NoteData, TreeItem, assertIsFolder, isFolder, isNote, isResource } from './types';
+import FolderRecord from './FolderRecord';
 import { extractResourceUrls } from '@joplin/lib/urlUtils';
-import ResourceRecord from './model/ResourceRecord';
+import ResourceRecord from './ResourceRecord';
 
 interface ClientData {
 	email: string;
